refactor(health): extract internal/health; webui drops tui import
- New internal/health package owns CheckStatus, CheckResult,
  DataProvider interface, and DefaultDataProvider with all 9
  built-in checks. Status enum: StatusOK/Warn/Err/Checking.
- internal/webui/handlers_health.go imports internal/health
  instead of internal/tui. Cross-presentation coupling removed.
- internal/tui/health_provider.go is now a thin shim of type
  aliases (HealthCheckStatus, HealthCheckResultMsg,
  HealthDataProvider, DefaultHealthDataProvider) and a
  NewDefaultHealthDataProvider constructor that delegates to the
  health package. All existing TUI consumers (health_list.go,
  health_detail.go, content_health.go, app.go, tests) keep their
  identifiers unchanged via aliasing.
- Verified: go list -deps internal/webui no longer contains
  internal/tui.

Partial of #1497 (subset 3.3).
